+++ CHANGELOG.md
@@ -2,5 +2,9 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2020.0.1 - 2020/10/08
+- Bugfix in md formatter
+- Overview table for md and ansi
+
 ## 2020 - 2020/10/08
 - First release